temp/SuntynAI-1/temp/SuntynAI/requirements_clean.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
werkzeug==3.1.3
gunicorn==23.0.0
pypdf2==3.0.1
pillow==11.3.0